+++ plugins/CoreUpdater/Commands/Update.php
@@ -16,6 +16,7 @@ use Piwik\Plugins\UserCountry\LocationProvider;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
  * @package CloudAdmin
@@ -28,33 +29,51 @@ class Update extends ConsoleCommand
 
         $this->setDescription('Triggers upgrades. Use it after Piwik core or any plugin files have been updated.');
 
-        $this->addOption('dry-run', null, InputOption::VALUE_NONE, 'Only prints out the SQL requests that would be executed during the upgrade');
+        $this->addOption('yes', null, InputOption::VALUE_NONE, 'Directly execute the update without asking for confirmation');
     }
 
     /**
-     * Execute command like: ./console core:update --dry-run
+     * Execute command like: ./console core:update --yes
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->executeClearCaches();
 
-        $doDryRun = (bool) $input->getOption('dry-run');
+        $yes = $input->getOption('yes');
 
         try {
-            $this->makeUpdate($input, $output, $doDryRun);
+            $this->makeUpdate($input, $output, true);
 
-            if (!$doDryRun) {
+            if (!$yes) {
+                $yes = $this->askForUpdateConfirmation($input, $output);
+            }
+
+            if ($yes) {
+                $output->writeln("\nStarting the database upgrade process now. This may take a while, so please be patient.");
+
+                $this->makeUpdate($input, $output, false);
+
                 $this->writeSuccessMessage($output, array("Piwik has been successfully updated!"));
+            } else {
+                $this->writeSuccessMessage($output, array('Database upgrade not executed.'));
             }
 
         } catch(NoUpdatesFoundException $e) {
             // Do not fail if no updates were found
-            $output->writeln("<info>".$e->getMessage()."</info>");
+            $this->writeSuccessMessage($output, array($e->getMessage()));
         } catch (\Exception $e) {
             // Fail in case of any other error during upgrade
             $output->writeln("<error>" . $e->getMessage() . "</error>");
             throw $e;
         }
+    }
+
+    private function askForUpdateConfirmation(InputInterface $input, OutputInterface $output)
+    {
+        $helper   = $this->getHelper('question');
+        $question = new ConfirmationQuestion('<comment>A database upgrade is required. Execute update? (y/N) </comment>', false);
+
+        return $helper->ask($input, $output, $question);
     }
 
     protected function executeClearCaches()
@@ -67,6 +86,8 @@ class Update extends ConsoleCommand
         $this->checkAllRequiredOptionsAreNotEmpty($input);
 
         $updateController = new Controller();
-        echo $updateController->runUpdaterAndExit($doDryRun);
+        $content = $updateController->runUpdaterAndExit($doDryRun);
+
+        $output->writeln($content);
     }
 }